# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Endometriosis Triage & Management Application API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List all patients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
        - name: triageLevel
          in: query
          schema:
            type: string
        - name: pathway
          in: query
          schema:
            type: string
        - name: carePathwayState
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Created patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get a patient by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Not found
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientUpdate"
      responses:
        "200":
          description: Updated patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    delete:
      operationId: deletePatient
      tags: [patients]
      summary: Delete a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /patients/{id}/surgeries:
    get:
      operationId: listPatientSurgeries
      tags: [surgeries]
      summary: List surgeries for a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of surgeries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Surgery"
    post:
      operationId: createSurgery
      tags: [surgeries]
      summary: Create a surgery record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SurgeryInput"
      responses:
        "201":
          description: Created surgery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Surgery"
  /surgeries/{id}:
    get:
      operationId: getSurgery
      tags: [surgeries]
      summary: Get a surgery record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Surgery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Surgery"
    patch:
      operationId: updateSurgery
      tags: [surgeries]
      summary: Update a surgery record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SurgeryUpdate"
      responses:
        "200":
          description: Updated surgery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Surgery"
...
  /medications:
    get:
      operationId: listMedications
      tags: [medications]
      summary: List all reference medications
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
        - name: tier
          in: query
          schema:
            type: string
        - name: formularyStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of medications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medication"
  /medications/{id}:
    get:
      operationId: getMedication
      tags: [medications]
      summary: Get a medication by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
...
components:
  schemas:
...
    Patient:
      type: object
      required: [id, firstName, lastName, dateOfBirth, carePathwayState, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        currentStage:
          type: ["string", "null"]
          description: "Endometriosis stage: Stage I, II, III, IV"
        triageLevel:
          type: ["string", "null"]
          description: "urgent, high, moderate, routine"
        lastAssessmentDate:
          type: ["string", "null"]
        carePathwayState:
          type: string
          description: "referral_received, clinic_review, imaging_completed, triage_decision, medication_started, waiting_list_added, mdt_discussed, surgery_scheduled, surgery_completed, post_op_review, follow_up_ongoing, discharged, long_term_management"
        currentPathway:
          type: ["string", "null"]
          description: "medical, surgery_general, surgery_specialist, combined"
        bsgeCentre:
          type: ["string", "null"]
        referralSource:
          type: ["string", "null"]
          description: "gp, self, a_e, fertility_clinic, other"
        referralDate:
          type: ["string", "null"]
        fertilityPriority:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    PatientInput:
      type: object
      required: [firstName, lastName, dateOfBirth]
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        dateOfBirth:
          type: string
        email:
          type: string
        phone:
          type: string
        notes:
          type: string
        referralSource:
          type: string
        referralDate:
          type: string
        fertilityPriority:
          type: boolean
    PatientUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
        email:
          type: string
        phone:
          type: string
        currentStage:
          type: string
        triageLevel:
          type: string
        carePathwayState:
          type: string
        currentPathway:
          type: string
        bsgeCentre:
          type: string
        referralSource:
          type: string
        referralDate:
          type: string
        fertilityPriority:
          type: boolean
        notes:
          type: string
...
    Surgery:
      type: object
      required: [id, patientId, procedureType, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        plannedDate:
          type: ["string", "null"]
        actualDate:
          type: ["string", "null"]
        surgeon:
          type: ["string", "null"]
        surgeonGrade:
          type: ["string", "null"]
        centre:
          type: ["string", "null"]
        bsgeAccredited:
          type: boolean
        procedureType:
          type: string
        procedureDetails:
          type: array
          items:
            type: string
        rafsStage:
          type: ["string", "null"]
        enzianScore:
          type: ["string", "null"]
        locations:
          type: array
          items:
            type: string
        endometriomaPresent:
          type: boolean
        endometriomaRight:
          type: boolean
        endometriomaLeft:
          type: boolean
        endometriomaRightSize:
          type: ["string", "null"]
        endometriomaLeftSize:
          type: ["string", "null"]
        deepEndometriosis:
          type: boolean
        deepEndometriosisLocation:
          type: ["string", "null"]
        adhesions:
          type: boolean
        adhesionsSeverity:
          type: ["string", "null"]
        complications:
          type: boolean
        complicationsDetails:
          type: ["string", "null"]
        estimatedBloodLoss:
          type: ["string", "null"]
        operativeTime:
          type: ["string", "null"]
        histologyConfirmed:
          type: boolean
        histologyDetails:
          type: ["string", "null"]
        postOpHormonalPlan:
          type: ["string", "null"]
        followUpRequired:
          type: boolean
        followUpWeeks:
          type: integer
        outcome:
          type: ["string", "null"]
        consentObtained:
          type: boolean
        consentDate:
          type: ["string", "null"]
        mdtDiscussed:
          type: boolean
        mdtDate:
          type: ["string", "null"]
        preOpNotes:
          type: ["string", "null"]
        intraOpNotes:
          type: ["string", "null"]
        postOpNotes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    SurgeryInput:
      type: object
      required: [patientId, procedureType]
      properties:
        patientId:
          type: integer
        plannedDate:
          type: string
        actualDate:
          type: string
        surgeon:
          type: string
        surgeonGrade:
          type: string
        centre:
          type: string
        bsgeAccredited:
          type: boolean
        procedureType:
          type: string
        procedureDetails:
          type: array
          items:
            type: string
        rafsStage:
          type: string
        enzianScore:
          type: string
        locations:
          type: array
          items:
            type: string
        endometriomaPresent:
          type: boolean
        endometriomaRight:
          type: boolean
        endometriomaLeft:
          type: boolean
        endometriomaRightSize:
          type: string
        endometriomaLeftSize:
          type: string
        deepEndometriosis:
          type: boolean
        deepEndometriosisLocation:
          type: string
        adhesions:
          type: boolean
        adhesionsSeverity:
          type: string
        complications:
          type: boolean
        complicationsDetails:
          type: string
        estimatedBloodLoss:
          type: string
        operativeTime:
          type: string
        histologyConfirmed:
          type: boolean
        histologyDetails:
          type: string
        postOpHormonalPlan:
          type: string
        followUpRequired:
          type: boolean
        followUpWeeks:
          type: integer
        outcome:
          type: string
        consentObtained:
          type: boolean
        consentDate:
          type: string
        mdtDiscussed:
          type: boolean
        mdtDate:
          type: string
        preOpNotes:
          type: string
        intraOpNotes:
          type: string
        postOpNotes:
          type: string
    SurgeryUpdate:
      type: object
      properties:
        plannedDate:
          type: string
        actualDate:
          type: string
        surgeon:
          type: string
        surgeonGrade:
          type: string
        centre:
          type: string
        bsgeAccredited:
          type: boolean
        procedureType:
          type: string
        procedureDetails:
          type: array
          items:
            type: string
        rafsStage:
          type: string
        enzianScore:
          type: string
        locations:
          type: array
          items:
            type: string
        endometriomaPresent:
          type: boolean
        endometriomaRight:
          type: boolean
        endometriomaLeft:
          type: boolean
        endometriomaRightSize:
          type: string
        endometriomaLeftSize:
          type: string
        deepEndometriosis:
          type: boolean
        deepEndometriosisLocation:
          type: string
        adhesions:
          type: boolean
        adhesionsSeverity:
          type: string
        complications:
          type: boolean
        complicationsDetails:
          type: string
        estimatedBloodLoss:
          type: string
        operativeTime:
          type: string
        histologyConfirmed:
          type: boolean
        histologyDetails:
          type: string
        postOpHormonalPlan:
          type: string
        followUpRequired:
          type: boolean
        followUpWeeks:
          type: integer
        outcome:
          type: string
        consentObtained:
          type: boolean
        consentDate:
          type: string
        mdtDiscussed:
          type: boolean
        mdtDate:
          type: string
        preOpNotes:
          type: string
        intraOpNotes:
          type: string
        postOpNotes:
          type: string
...
    Medication:
      type: object
      required: [id, name, category, mechanism, indications]
      properties:
        id:
          type: integer
        name:
          type: string
        genericName:
          type: ["string", "null"]
        category:
          type: string
          description: "hormonal, pain-relief, gnrh-agonist, gnrh-antagonist, nsaid, progestin, ocp, other"
        mechanism:
          type: string
        indications:
          type: string
        contraindications:
          type: ["string", "null"]
        sideEffects:
          type: ["string", "null"]
        dosage:
          type: ["string", "null"]
        approvedStages:
          type: array
          items:
            type: string
        evidenceLevel:
          type: ["string", "null"]
          description: "Level A, B, C"
        tier:
          type: string
          description: "tier1, tier2, tier3, analgesia"
        formularyStatus:
          type: string
          description: "green, amber, red"
        formularyNotes:
          type: ["string", "null"]
        niceTa:
          type: ["string", "null"]
        niceApproved:
          type: boolean
        notes:
          type: ["string", "null"]
...