# lib/api-zod/src/generated/types/surgery.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Endometriosis Triage & Management Application API
 * OpenAPI spec version: 0.1.0
 */

export interface Surgery {
  id: number;
  patientId: number;
  /** @nullable */
  plannedDate?: string | null;
  /** @nullable */
  actualDate?: string | null;
  /** @nullable */
  surgeon?: string | null;
  /** @nullable */
  surgeonGrade?: string | null;
  /** @nullable */
  centre?: string | null;
  bsgeAccredited?: boolean;
  procedureType: string;
  procedureDetails?: string[];
  /** @nullable */
  rafsStage?: string | null;
  /** @nullable */
  enzianScore?: string | null;
  locations?: string[];
  endometriomaPresent?: boolean;
  endometriomaRight?: boolean;
  endometriomaLeft?: boolean;
  /** @nullable */
  endometriomaRightSize?: string | null;
  /** @nullable */
  endometriomaLeftSize?: string | null;
  deepEndometriosis?: boolean;
  /** @nullable */
  deepEndometriosisLocation?: string | null;
  adhesions?: boolean;
  /** @nullable */
  adhesionsSeverity?: string | null;
  complications?: boolean;
  /** @nullable */
  complicationsDetails?: string | null;
  /** @nullable */
  estimatedBloodLoss?: string | null;
  /** @nullable */
  operativeTime?: string | null;
  histologyConfirmed?: boolean;
  /** @nullable */
  histologyDetails?: string | null;
  /** @nullable */
  postOpHormonalPlan?: string | null;
  followUpRequired?: boolean;
  followUpWeeks?: number;
  /** @nullable */
  outcome?: string | null;
  consentObtained?: boolean;
  /** @nullable */
  consentDate?: string | null;
  mdtDiscussed?: boolean;
  /** @nullable */
  mdtDate?: string | null;
  /** @nullable */
  preOpNotes?: string | null;
  /** @nullable */
  intraOpNotes?: string | null;
  /** @nullable */
  postOpNotes?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}