# lib/api-zod/src/generated/types/surgeryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Endometriosis Triage & Management Application API
 * OpenAPI spec version: 0.1.0
 */

export interface SurgeryInput {
  patientId: number;
  plannedDate?: string;
  actualDate?: string;
  surgeon?: string;
  surgeonGrade?: string;
  centre?: string;
  bsgeAccredited?: boolean;
  procedureType: string;
  procedureDetails?: string[];
  rafsStage?: string;
  enzianScore?: string;
  locations?: string[];
  endometriomaPresent?: boolean;
  endometriomaRight?: boolean;
  endometriomaLeft?: boolean;
  endometriomaRightSize?: string;
  endometriomaLeftSize?: string;
  deepEndometriosis?: boolean;
  deepEndometriosisLocation?: string;
  adhesions?: boolean;
  adhesionsSeverity?: string;
  complications?: boolean;
  complicationsDetails?: string;
  estimatedBloodLoss?: string;
  operativeTime?: string;
  histologyConfirmed?: boolean;
  histologyDetails?: string;
  postOpHormonalPlan?: string;
  followUpRequired?: boolean;
  followUpWeeks?: number;
  outcome?: string;
  consentObtained?: boolean;
  consentDate?: string;
  mdtDiscussed?: boolean;
  mdtDate?: string;
  preOpNotes?: string;
  intraOpNotes?: string;
  postOpNotes?: string;
}